requirements.txt
requests==2.28.1
rich==12.5.1
python-dotenv==1.0.0
